+++ default.js
@@ -87,6 +87,7 @@ module.exports = {
     'no-shadow-restricted-names': 2,
     'no-undefined': 2,
     'no-undef-init': 2,
+    'no-unused-vars': 2,
     'no-use-before-define': 0,
 
     // Node.js and CommonJS